#275 reorganize idea tests
